BP-1328: Add TTL to robot parameters
Parameters on Robots now have a new attribute called 'TTL'. If set,
whenever the Value of that parameter is updated, the system will
record the current timestamp. Then, when the Value gets read,
if the Value was recorded too long ago, it'll return None instead.

TTL is an integer, representing a number of seconds.