BP-1365: Fix KeyError when running some flows
When parsing the flow parameters, the mechanism would read the
default parameter from the subflow, despite it being overridden
on the top flow. This PR makes it that the default is only parsed
if there's no expression overridding it.

Also, we replace the custom parser with a more robust one written
using Lark (https://pypi.org/project/lark/), which allows for
clearer tracebacks, since the expressions become part of the
Python execution stack.