Check existance after getting lock
Imagine this function is called twice when the class still doesn't exist. Both pass the if and get to the lock. One of them will get the lock first, the other will wait, but when the first releases the lock, the second will grab the lock and create a new instance, replacing the one created by the first.

Co-authored-by: andreparames <andreparames@users.noreply.github.com>